facilitators/xrp-server/requirements.txt
# XRP X402 Facilitator Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
xrpl-py>=2.6.0
httpx>=0.25.0
pydantic>=2.4.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.0
requests>=2.31.0